style: Support calc in cursor.
Bug: 1453258
MozReview-Commit-ID: 1ZzxkYboWZg
